+++ src/main/java/org/folio/search/service/setter/callnumber/DefaultCallNumberShelvingOrderFieldProcessor.java
@@ -1,0 +1,13 @@
+package org.folio.search.service.setter.callnumber;
+
+import org.folio.search.domain.dto.ShelvingOrderAlgorithmType;
+import org.folio.search.utils.ShelvingOrderCalculationHelper;
+import org.springframework.stereotype.Component;
+
+@Component
+public class DefaultCallNumberShelvingOrderFieldProcessor extends CallNumberShelvingOrderFieldProcessor {
+
+  protected DefaultCallNumberShelvingOrderFieldProcessor() {
+    super(number -> ShelvingOrderCalculationHelper.calculate(number, ShelvingOrderAlgorithmType.DEFAULT));
+  }
+}