+++ src/main/java/org/folio/search/service/setter/callnumber/DeweyCallNumberShelvingOrderFieldProcessor.java
@@ -1,13 +1,12 @@
-package org.folio.search.service.setter.classification;
+package org.folio.search.service.setter.callnumber;
 
 import org.folio.search.domain.dto.ShelvingOrderAlgorithmType;
 import org.folio.search.utils.ShelvingOrderCalculationHelper;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DeweyClassificationShelvingOrderFieldProcessor extends ClassificationShelvingOrderFieldProcessor {
-
-  protected DeweyClassificationShelvingOrderFieldProcessor() {
+public class DeweyCallNumberShelvingOrderFieldProcessor extends CallNumberShelvingOrderFieldProcessor {
+  protected DeweyCallNumberShelvingOrderFieldProcessor() {
     super(number -> ShelvingOrderCalculationHelper.calculate(number, ShelvingOrderAlgorithmType.DEWEY));
   }
 }